Split domain quiz data structure into domain/REST
With conversions between them, and tests for those conversions.

As part of this, some intermediated maps and arrays are now in a
QuizCache structure, used to fill the extras in the REST data
structures, to avoid putting too much in the actual REST data
structures.

We now have simpler REST (API) structs and it's easier to see what is in
the REST API, what is in the domain objects, and what is in the database
(dto).