GitHub CI action: Add google-github-actions/auth step
So the subsequent google-github-actions/setup-gcloud@v3 can succeed.
This has been broken since I added the use of that
google-github-actions/setup-gcloud in
https://github.com/murraycu/go-bigoquiz-server/commit/3d7e02fae1624202966b0abf683b114682ca7892

Before this it helpfully warns like so:
"
Warning: The gcloud CLI is not authenticated (or it is not installed). Authenticate by adding the "google-github-actions/auth" step prior this one.
"

Add a google-github-actions/auth@v2 step using a "project_id" value of "bigoquiz", and a "workload_identity_provider" value generated as per the  instructions at:
https://github.com/google-github-actions/auth?tab=readme-ov-file#preferred-direct-workload-identity-federation:

I ran these commands:

  $ export PROJECT_ID=bigoquiz
  $ gcloud iam workload-identity-pools create "github"   --project="${PROJECT_ID}"   --location="global"   --display-name="GitHub Actions Pool"
  $ gcloud iam workload-identity-pools describe "github"   --project="${PROJECT_ID}"   --location="global"   --format="value(name)"

to get this full ID of the Workload Identity Pool (though later steps here don't seem to use this ID directly)
  projects/710014864852/locations/global/workloadIdentityPools/github

  $ export GITHUB_ORG="murraycu"
  $ gcloud iam workload-identity-pools providers create-oidc "my-repo"   --project="${PROJECT_ID}"   --location="global"   --workload-identity-pool="github"   --display-name="My GitHub repo Provider"   --attribute-mapping="google.subject=assertion.sub,attribute.actor=assertion.actor,attribute.repository=assertion.repository,attribute.repository_owner=assertion.repository_owner"   --attribute-condition="assertion.repository_owner == '${GITHUB_ORG}'"   --issuer-uri="https://token.actions.githubusercontent.com"
  $ gcloud iam workload-identity-pools providers describe "my-repo"   --project="${PROJECT_ID}"   --location="global"   --workload-identity-pool="github"   --format="value(name)"

to get this "Workload Identity Provider resource name" (the workload_identity_provider then used in the workflow .yml file):
  projects/710014864852/locations/global/workloadIdentityPools/github/providers/my-repo